+++ mongoengine/signals.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 
-__all__ = ['pre_init', 'post_init', 'pre_save_validation', 'pre_save',
+__all__ = ['pre_init', 'post_init', 'pre_save', 'pre_save_post_validation',
            'post_save', 'pre_delete', 'post_delete']
 
 signals_available = False
@@ -38,8 +38,8 @@ _signals = Namespace()
 
 pre_init = _signals.signal('pre_init')
 post_init = _signals.signal('post_init')
-pre_save_validation = _signals.signal('pre_save_validation')
 pre_save = _signals.signal('pre_save')
+pre_save_post_validation = _signals.signal('pre_save_post_validation')
 post_save = _signals.signal('post_save')
 pre_delete = _signals.signal('pre_delete')
 post_delete = _signals.signal('post_delete')